requirements.txt
opencv-python
numpy
fer
tensorflow
pillow
pandas
matplotlib
scikit-learn